requirements.txt
beautifulsoup4
capmonster_python
emoji
flask
jaconv
markovify
mecab
nuitka
pillow
pyarmor
pypresence
requests
tls_client
unidic-lite
websocket-client